refactored blueprints into dedicated directory
