+++ culturemesh/blueprints/user_home/templates/networks.html
